feat: pushing back to dev container for checking - moved coordinator outside
