+++ .coveragerc
@@ -508,6 +508,7 @@ omit =
     homeassistant/components/insteon/schemas.py
     homeassistant/components/insteon/switch.py
     homeassistant/components/insteon/utils.py
+    homeassistant/components/intellifire/__init__.py
     homeassistant/components/intellifire/binary_sensor.py
     homeassistant/components/incomfort/*
     homeassistant/components/intesishome/*
